backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyjwt==2.10.1
werkzeug==3.1.3
gunicorn==23.0.0
